PLT-1042 allow cancel pending post
