+++ webapp/i18n/en.json
@@ -1063,6 +1063,8 @@
   "password_send.link": "<p>A password reset link has been sent to <b>{email}</b></p>",
   "password_send.reset": "Reset my password",
   "password_send.title": "Password Reset",
+  "pending_post_actions.retry": "Retry",
+  "pending_post_actions.cancel": "Cancel",
   "permalink.error.access": "Permalink belongs to a channel you do not have access to",
   "post_attachment.collapse": "▲ collapse text",
   "post_attachment.more": "▼ read more",
@@ -1070,7 +1072,6 @@
   "post_body.deleted": "(message deleted)",
   "post_body.plusMore": " plus {count} other files",
   "post_body.plusOne": " plus 1 other file",
-  "post_body.retry": "Retry",
   "post_delete.notPosted": "Comment could not be posted",
   "post_delete.okay": "Okay",
   "post_delete.someone": "Someone deleted the message on which you tried to post a comment.",
@@ -1120,7 +1121,6 @@
   "rhs_comment.del": "Delete",
   "rhs_comment.edit": "Edit",
   "rhs_comment.permalink": "Permalink",
-  "rhs_comment.retry": "Retry",
   "rhs_header.details": "Message Details",
   "rhs_root.del": "Delete",
   "rhs_root.direct": "Direct Message",
