+++ webapp/i18n/es.json
@@ -1063,6 +1063,8 @@
   "password_send.link": "<p>Se ha enviado un enlace para restablecer la contraseña a <b>{email}</b></p>",
   "password_send.reset": "Restablecer mi contraseña",
   "password_send.title": "Restablecer Contraseña",
+  "pending_post_actions.retry": "Reintentar",
+  "pending_post_actions.cancel": "Cancelar",
   "permalink.error.access": "El enlace permanente pertenece a un canal al cual no tienes acceso",
   "post_attachment.collapse": "▲ colapsar texto",
   "post_attachment.more": "▼ leer más",
@@ -1070,7 +1072,6 @@
   "post_body.deleted": "(mensaje eliminado)",
   "post_body.plusMore": " más {count} otros archivos",
   "post_body.plusOne": " más 1 archivo",
-  "post_body.retry": "Reintentar",
   "post_delete.notPosted": "No se pudo enviar el comentario",
   "post_delete.okay": "Ok",
   "post_delete.someone": "Alguien borró el mensaje que querías comentar.",
@@ -1120,7 +1121,6 @@
   "rhs_comment.del": "Borrar",
   "rhs_comment.edit": "Editar",
   "rhs_comment.permalink": "Enlace permanente",
-  "rhs_comment.retry": "Reintentar",
   "rhs_header.details": "Detalles del Mensaje",
   "rhs_root.del": "Borrar",
   "rhs_root.direct": "Mensaje Directo",
