+++ webapp/i18n/fr.json
@@ -1063,6 +1063,8 @@
   "password_send.link": "<p>Un lien de réinitialisation de mot de passe a été envoyé à <b>{email}</b></p>",
   "password_send.reset": "Réinitialiser mon mot de passe",
   "password_send.title": "Réinitialisation mot de passe",
+  "pending_post_actions.retry": "Réessayer",
+  "pending_post_actions.cancel": "Annuler",
   "permalink.error.access": "Permalink belongs to a channel you do not have access to",
   "post_attachment.collapse": "▲ réduire le texte",
   "post_attachment.more": "▼ lire la suite",
@@ -1070,7 +1072,6 @@
   "post_body.deleted": "(message supprimé)",
   "post_body.plusMore": " et {count} autres fichiers",
   "post_body.plusOne": " et 1 autre fichier",
-  "post_body.retry": "Réessayer",
   "post_delete.notPosted": "Le commentaire n'a pu être envoyé",
   "post_delete.okay": "Ok",
   "post_delete.someone": "Quelqu'un a supprimé le message sur lequel vous tentiez d'envoyer un commentaire.",
@@ -1120,7 +1121,6 @@
   "rhs_comment.del": "Supprimer",
   "rhs_comment.edit": "Éditer",
   "rhs_comment.permalink": "Lien permanent",
-  "rhs_comment.retry": "Réessayer",
   "rhs_header.details": "Détails du message",
   "rhs_root.del": "Supprimer",
   "rhs_root.direct": "Messages privés",
