+++ webapp/i18n/ja.json
@@ -1063,6 +1063,8 @@
   "password_send.link": "<p>パスワード初期化リンクが<b>{email}</b>に送信されました</p>",
   "password_send.reset": "自分のパスワードを初期化する",
   "password_send.title": "パスワードの初期化",
+  "pending_post_actions.retry": "再試行する",
+  "pending_post_actions.cancel": "キャンセル",
   "permalink.error.access": "アクセスできないチャンネルのパーマリンクです",
   "post_attachment.collapse": "▲テキストをたたむ",
   "post_attachment.more": "▼もっと読む",
@@ -1070,7 +1072,6 @@
   "post_body.deleted": "(メッセージは削除されています)",
   "post_body.plusMore": " {count}の他のファイルを追加する",
   "post_body.plusOne": " 1つの他のファイルを追加する",
-  "post_body.retry": "再試行する",
   "post_delete.notPosted": "コメントが投稿できませんでした",
   "post_delete.okay": "削除する",
   "post_delete.someone": "あなたがコメントしようとしたメッセージは削除されています。",
@@ -1120,7 +1121,6 @@
   "rhs_comment.del": "削除する",
   "rhs_comment.edit": "編集する",
   "rhs_comment.permalink": "パーマリンク",
-  "rhs_comment.retry": "再試行する",
   "rhs_header.details": "メッセージの詳細",
   "rhs_root.del": "削除する",
   "rhs_root.direct": "ダイレクトメッセージ",
