+++ webapp/i18n/pt.json
@@ -1063,6 +1063,8 @@
   "password_send.link": "<p>Um link para resetar senha foi enviado para <b>{email}</b></p>",
   "password_send.reset": "Resetar minha senha",
   "password_send.title": "Resetar Senha",
+  "pending_post_actions.retry": "Tentar novamente",
+  "pending_post_actions.cancel": "Cancelar",
   "permalink.error.access": "Permalink pertence a um canal que você não tem acesso",
   "post_attachment.collapse": "▲ recolher texto",
   "post_attachment.more": "▼ leia mais",
@@ -1070,7 +1072,6 @@
   "post_body.deleted": "(mensagem deletada)",
   "post_body.plusMore": " mais {count} outros arquivos",
   "post_body.plusOne": " mais 1 outro arquivo",
-  "post_body.retry": "Tentar novamente",
   "post_delete.notPosted": "Comentário não pode ser postado",
   "post_delete.okay": "Ok",
   "post_delete.someone": "Alguém deletou a mensagem no qual você tentou postar um comentário.",
@@ -1120,7 +1121,6 @@
   "rhs_comment.del": "Deletar",
   "rhs_comment.edit": "Editar",
   "rhs_comment.permalink": "Permalink",
-  "rhs_comment.retry": "Tentar novamente",
   "rhs_header.details": "Detalhes da Mensagem",
   "rhs_root.del": "Deletar",
   "rhs_root.direct": "Mensagem Direta",
